\name{BayesianIntro-package}
\alias{BayesianIntro-package}
\alias{BayesianIntro}
\docType{package}
\title{
Introduction To Bayesian Computation Data
}
\description{
Houses data for an introduction to Bayesian computation text. 
}
\details{
\tabular{ll}{
Package: \tab BayesianIntro\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-13\cr
License: \tab What license is it under?\cr
}
}
\author{
Kevin Tenenbaum

Maintainer: Kevin Tenenbaum <ktenenbaum@middlebury.edu>
}
\keyword{ package }