% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BayesianIntro.R
\docType{data}
\name{OBPData}
\alias{OBPData}
\title{OBPData.}
\format{A data frame with 4 variables: \code{playerID}, \code{Name}, \code{PA} and
  \code{OBP}}
\usage{
OBPData
}
\description{
On Base Percentage Data
}
\keyword{datasets}

