% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BayesianIntro.R
\docType{data}
\name{pitchfx}
\alias{pitchfx}
\title{pitchfx.}
\format{A data frame with seven variables: \code{X}, \code{pitcher},
  \code{name}, \code{velocity}, \code{horizMovement}, \code{vertMovement}, and \code{KPCT}}
\usage{
pitchfx
}
\description{
Pitchfx data
}
\keyword{datasets}

